+++ tests/md5.rs
@@ -1,56 +1,59 @@
-use chksum_hash::md5;
+use chksum_hash::{default, hash, MD5};
 
 #[test]
 fn hash_empty() {
-    let digest = md5::hash("").to_hex_lowercase();
+    let digest = hash::<MD5, _>("").to_hex_lowercase();
     assert_eq!(digest, "d41d8cd98f00b204e9800998ecf8427e");
 
-    let digest = md5::hash(b"").to_hex_lowercase();
+    let digest = hash::<MD5, _>(b"").to_hex_lowercase();
     assert_eq!(digest, "d41d8cd98f00b204e9800998ecf8427e");
 
-    let digest = md5::hash(b"".to_vec()).to_hex_lowercase();
+    let digest = hash::<MD5, _>(b"".to_vec()).to_hex_lowercase();
     assert_eq!(digest, "d41d8cd98f00b204e9800998ecf8427e");
 }
 
 #[test]
 fn new_empty() {
-    let digest = md5::new().digest().to_hex_lowercase();
+    let digest = default::<MD5>().digest().to_hex_lowercase();
     assert_eq!(digest, "d41d8cd98f00b204e9800998ecf8427e");
 
-    let digest = md5::new().update("").digest().to_hex_lowercase();
+    let digest = default::<MD5>().update("").digest().to_hex_lowercase();
     assert_eq!(digest, "d41d8cd98f00b204e9800998ecf8427e");
 
-    let digest = md5::new().update(b"").digest().to_hex_lowercase();
+    let digest = default::<MD5>().update(b"").digest().to_hex_lowercase();
     assert_eq!(digest, "d41d8cd98f00b204e9800998ecf8427e");
 
-    let digest = md5::new().update(b"".to_vec()).digest().to_hex_lowercase();
+    let digest = default::<MD5>().update(b"".to_vec()).digest().to_hex_lowercase();
     assert_eq!(digest, "d41d8cd98f00b204e9800998ecf8427e");
 }
 
 #[test]
 fn hash_hello_world() {
-    let digest = md5::hash("Hello World!").to_hex_lowercase();
+    let digest = hash::<MD5, _>("Hello World!").to_hex_lowercase();
     assert_eq!(digest, "ed076287532e86365e841e92bfc50d8c");
 
-    let digest = md5::hash(b"Hello World!").to_hex_lowercase();
+    let digest = hash::<MD5, _>(b"Hello World!").to_hex_lowercase();
     assert_eq!(digest, "ed076287532e86365e841e92bfc50d8c");
 
-    let digest = md5::hash(b"Hello World!".to_vec()).to_hex_lowercase();
+    let digest = hash::<MD5, _>(b"Hello World!".to_vec()).to_hex_lowercase();
     assert_eq!(digest, "ed076287532e86365e841e92bfc50d8c");
 }
 
 #[test]
 fn new_hello_world() {
-    let digest = md5::new().update("Hello World!").digest().to_hex_lowercase();
+    let digest = default::<MD5>().update("Hello World!").digest().to_hex_lowercase();
     assert_eq!(digest, "ed076287532e86365e841e92bfc50d8c");
 
-    let digest = md5::new().update(b"Hello World!").digest().to_hex_lowercase();
+    let digest = default::<MD5>().update(b"Hello World!").digest().to_hex_lowercase();
     assert_eq!(digest, "ed076287532e86365e841e92bfc50d8c");
 
-    let digest = md5::new().update(b"Hello World!".to_vec()).digest().to_hex_lowercase();
+    let digest = default::<MD5>()
+        .update(b"Hello World!".to_vec())
+        .digest()
+        .to_hex_lowercase();
     assert_eq!(digest, "ed076287532e86365e841e92bfc50d8c");
 
-    let digest = md5::new()
+    let digest = default::<MD5>()
         .update("Hello")
         .update(" ")
         .update("World!")
@@ -58,7 +61,7 @@ fn new_hello_world() {
         .to_hex_lowercase();
     assert_eq!(digest, "ed076287532e86365e841e92bfc50d8c");
 
-    let digest = md5::new()
+    let digest = default::<MD5>()
         .update(b"Hello")
         .update(b" ")
         .update(b"World!")
@@ -66,7 +69,7 @@ fn new_hello_world() {
         .to_hex_lowercase();
     assert_eq!(digest, "ed076287532e86365e841e92bfc50d8c");
 
-    let digest = md5::new()
+    let digest = default::<MD5>()
         .update(b"Hello".to_vec())
         .update(b" ".to_vec())
         .update(b"World!".to_vec())
@@ -74,7 +77,7 @@ fn new_hello_world() {
         .to_hex_lowercase();
     assert_eq!(digest, "ed076287532e86365e841e92bfc50d8c");
 
-    let digest = md5::new()
+    let digest = default::<MD5>()
         .update("Hello")
         .update(b" ")
         .update(b"World!".to_vec())