+++ tests/sha2_256.rs
@@ -1,20 +1,20 @@
-use chksum_hash::sha2_256;
+use chksum_hash::{default, hash, SHA2_256};
 
 #[test]
 fn hash_empty() {
-    let digest = sha2_256::hash("").to_hex_lowercase();
+    let digest = hash::<SHA2_256, _>("").to_hex_lowercase();
     assert_eq!(
         digest,
         "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
     );
 
-    let digest = sha2_256::hash(b"").to_hex_lowercase();
+    let digest = hash::<SHA2_256, _>(b"").to_hex_lowercase();
     assert_eq!(
         digest,
         "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
     );
 
-    let digest = sha2_256::hash(b"".to_vec()).to_hex_lowercase();
+    let digest = hash::<SHA2_256, _>(b"".to_vec()).to_hex_lowercase();
     assert_eq!(
         digest,
         "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
@@ -23,25 +23,25 @@ fn hash_empty() {
 
 #[test]
 fn new_empty() {
-    let digest = sha2_256::new().digest().to_hex_lowercase();
+    let digest = default::<SHA2_256>().digest().to_hex_lowercase();
     assert_eq!(
         digest,
         "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
     );
 
-    let digest = sha2_256::new().update("").digest().to_hex_lowercase();
+    let digest = default::<SHA2_256>().update("").digest().to_hex_lowercase();
     assert_eq!(
         digest,
         "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
     );
 
-    let digest = sha2_256::new().update(b"").digest().to_hex_lowercase();
+    let digest = default::<SHA2_256>().update(b"").digest().to_hex_lowercase();
     assert_eq!(
         digest,
         "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
     );
 
-    let digest = sha2_256::new().update(b"".to_vec()).digest().to_hex_lowercase();
+    let digest = default::<SHA2_256>().update(b"".to_vec()).digest().to_hex_lowercase();
     assert_eq!(
         digest,
         "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
@@ -50,19 +50,19 @@ fn new_empty() {
 
 #[test]
 fn hash_hello_world() {
-    let digest = sha2_256::hash("Hello World!").to_hex_lowercase();
+    let digest = hash::<SHA2_256, _>("Hello World!").to_hex_lowercase();
     assert_eq!(
         digest,
         "7f83b1657ff1fc53b92dc18148a1d65dfc2d4b1fa3d677284addd200126d9069"
     );
 
-    let digest = sha2_256::hash(b"Hello World!").to_hex_lowercase();
+    let digest = hash::<SHA2_256, _>(b"Hello World!").to_hex_lowercase();
     assert_eq!(
         digest,
         "7f83b1657ff1fc53b92dc18148a1d65dfc2d4b1fa3d677284addd200126d9069"
     );
 
-    let digest = sha2_256::hash(b"Hello World!".to_vec()).to_hex_lowercase();
+    let digest = hash::<SHA2_256, _>(b"Hello World!".to_vec()).to_hex_lowercase();
     assert_eq!(
         digest,
         "7f83b1657ff1fc53b92dc18148a1d65dfc2d4b1fa3d677284addd200126d9069"
@@ -71,19 +71,22 @@ fn hash_hello_world() {
 
 #[test]
 fn new_hello_world() {
-    let digest = sha2_256::new().update("Hello World!").digest().to_hex_lowercase();
+    let digest = default::<SHA2_256>().update("Hello World!").digest().to_hex_lowercase();
     assert_eq!(
         digest,
         "7f83b1657ff1fc53b92dc18148a1d65dfc2d4b1fa3d677284addd200126d9069"
     );
 
-    let digest = sha2_256::new().update(b"Hello World!").digest().to_hex_lowercase();
+    let digest = default::<SHA2_256>()
+        .update(b"Hello World!")
+        .digest()
+        .to_hex_lowercase();
     assert_eq!(
         digest,
         "7f83b1657ff1fc53b92dc18148a1d65dfc2d4b1fa3d677284addd200126d9069"
     );
 
-    let digest = sha2_256::new()
+    let digest = default::<SHA2_256>()
         .update(b"Hello World!".to_vec())
         .digest()
         .to_hex_lowercase();
@@ -92,7 +95,7 @@ fn new_hello_world() {
         "7f83b1657ff1fc53b92dc18148a1d65dfc2d4b1fa3d677284addd200126d9069"
     );
 
-    let digest = sha2_256::new()
+    let digest = default::<SHA2_256>()
         .update("Hello")
         .update(" ")
         .update("World!")
@@ -103,7 +106,7 @@ fn new_hello_world() {
         "7f83b1657ff1fc53b92dc18148a1d65dfc2d4b1fa3d677284addd200126d9069"
     );
 
-    let digest = sha2_256::new()
+    let digest = default::<SHA2_256>()
         .update(b"Hello")
         .update(b" ")
         .update(b"World!")
@@ -114,7 +117,7 @@ fn new_hello_world() {
         "7f83b1657ff1fc53b92dc18148a1d65dfc2d4b1fa3d677284addd200126d9069"
     );
 
-    let digest = sha2_256::new()
+    let digest = default::<SHA2_256>()
         .update(b"Hello".to_vec())
         .update(b" ".to_vec())
         .update(b"World!".to_vec())
@@ -125,7 +128,7 @@ fn new_hello_world() {
         "7f83b1657ff1fc53b92dc18148a1d65dfc2d4b1fa3d677284addd200126d9069"
     );
 
-    let digest = sha2_256::new()
+    let digest = default::<SHA2_256>()
         .update("Hello")
         .update(b" ")
         .update(b"World!".to_vec())